+++ integration-tests/test_full_f144_functionality.py
@@ -9,7 +9,7 @@ from helpers.kafkahelpers import (
 )
 from datetime import datetime, timedelta
 from file_writer_control.WriteJob import WriteJob
-from helpers import full_file_path
+from helpers import build_relative_file_path
 from helpers.writer import (
     wait_start_job,
     wait_writers_available,
@@ -19,7 +19,7 @@ import numpy as np
 
 
 def test_f144(worker_pool, kafka_address, hdf_file_name="scal_output_file.nxs"):
-    file_path = full_file_path(hdf_file_name)
+    file_path = build_relative_file_path(hdf_file_name)
     wait_writers_available(worker_pool, nr_of=1, timeout=20)
     producer = create_producer(kafka_address)