+++ src/tests/MasterTests.cpp
@@ -77,8 +77,20 @@ public:
     UnderTest = std::make_unique<FileWriter::Master>(
         Config, std::move(TmpCmdHandler), std::move(TmpStatusReporter),
         Registrar);
-    if (std::filesystem::exists(StartCmd.Filename)) {
-      std::filesystem::remove(StartCmd.Filename);
+
+    std::filesystem::path FullFilePath = Config.getHDFOutputPrefix();
+    FullFilePath.append(StartCmd.Filename);
+    if (!StartCmd.Filename.empty() && std::filesystem::exists(FullFilePath)) {
+      std::filesystem::remove(FullFilePath);
+    }
+    FullFilePath =
+        std::filesystem::path(Config.getHDFOutputPrefix()) /
+        std::filesystem::path(StartCmdAbsoluteFilename).relative_path();
+    if (std::filesystem::exists(FullFilePath)) {
+      std::filesystem::remove(FullFilePath);
+    }
+    if (!std::filesystem::exists(FullFilePath.parent_path())) {
+      std::filesystem::create_directory(FullFilePath.parent_path());
     }
   }
   MainOpt Config;
@@ -87,19 +99,44 @@ public:
   StatusReporterStandIn *StatusReporter;
   std::unique_ptr<FileWriter::Master> UnderTest;
   time_point StartTime{system_clock::now()};
-  std::filesystem::path FilePath =
-      std::filesystem::current_path().append("file_name");
   Command::StartInfo StartCmd{"job_id",
-                              FilePath,
+                              "some_file_name",
                               R"({"nexus_structure":5})",
                               R"({"meta_data":54})",
                               StartTime,
                               StartTime + 50s,
                               "control_topic"};
+  std::string StartCmdAbsoluteFilename{
+      std::filesystem::temp_directory_path().append("some_file_name")};
+  Command::StartInfo StartCmdAbsolute{"job_id",
+                                      StartCmdAbsoluteFilename,
+                                      R"({"nexus_structure":5})",
+                                      R"({"meta_data":54})",
+                                      StartTime,
+                                      StartTime + 50s,
+                                      "control_topic"};
 };
 
 TEST_F(MasterTest, Init) {
   // Do nothing extra here, its all done in the SetUp()-function
+}
+
+TEST_F(MasterTest, DestinationFilenameFromRelativePath) {
+  REQUIRE_CALL(*StatusReporter,
+               useAlternativeStatusTopic(StartCmd.ControlTopic));
+  UnderTest->startWriting(StartCmd);
+  std::filesystem::path FullFilePath =
+      std::filesystem::path(Config.getHDFOutputPrefix()) / StartCmd.Filename;
+  EXPECT_EQ(UnderTest->getCurrentFilePath(), FullFilePath);
+}
+
+TEST_F(MasterTest, DestinationFilenameFromAbsolutePath) {
+  REQUIRE_CALL(*StatusReporter,
+               useAlternativeStatusTopic(StartCmdAbsolute.ControlTopic));
+  UnderTest->startWriting(StartCmdAbsolute);
+  std::filesystem::path FullFilePath = std::filesystem::path(
+      Config.getHDFOutputPrefix() + StartCmdAbsolute.Filename);
+  EXPECT_EQ(UnderTest->getCurrentFilePath(), FullFilePath);
 }
 
 TEST_F(MasterTest, StartWritingSuccess) {